+++ Libplanet.Explorer.Executable/Program.cs
@@ -17,6 +17,7 @@ using Libplanet.Explorer.Executable.Exceptions;
 using Libplanet.Explorer.Interfaces;
 using Libplanet.Explorer.Store;
 using Libplanet.Net;
+using Libplanet.Net.Protocols;
 using Libplanet.Store;
 using Libplanet.Store.Trie;
 using Libplanet.Tx;
@@ -229,6 +230,10 @@ If omitted (default) explorer only the local blockchain store.")]
                     // FIXME: The appProtocolVersion should be fixed properly.
                     var swarmOptions = new SwarmOptions
                     {
+                        BootstrapOptions = new BootstrapOptions
+                        {
+                            SeedPeers = options.Seeds.ToImmutableList(),
+                        },
                         TimeoutOptions = new TimeoutOptions
                         {
                             MaxTimeout = TimeSpan.FromSeconds(10),
@@ -261,7 +266,7 @@ If omitted (default) explorer only the local blockchain store.")]
                     {
                         await Task.WhenAll(
                             webHost.RunAsync(cts.Token),
-                            StartSwarmAsync(swarm, options.Seeds, cts.Token)
+                            StartSwarmAsync(swarm, cts.Token)
                         );
                     }
                     catch (OperationCanceledException)
@@ -355,7 +360,6 @@ If omitted (default) explorer only the local blockchain store.")]
 
         private static async Task StartSwarmAsync(
             Swarm<NullAction> swarm,
-            IEnumerable<Peer> seeds,
             CancellationToken cancellationToken)
         {
             if (swarm is null)
@@ -367,11 +371,7 @@ If omitted (default) explorer only the local blockchain store.")]
             try
             {
                 Console.Error.WriteLine("Bootstrapping.");
-                await swarm.BootstrapAsync(
-                    seeds,
-                    TimeSpan.FromMilliseconds(5000),
-                    cancellationToken: cancellationToken
-                );
+                await swarm.BootstrapAsync(cancellationToken);
             }
             catch (TimeoutException)
             {